safari-534.53-branch/LayoutTests/svg/dynamic-updates/SVGRectElement-svgdom-requiredFeatures-expected.txt

SVG 1.1 dynamic update tests

Check that SVGRectElement is initially displayed
PASS document.defaultView.getComputedStyle(rectElement, null).display is "inline"
Check that setting requiredFeatures to something invalid makes it not render
PASS document.defaultView.getComputedStyle(rectElement, null).display is ""
Check that setting requiredFeatures to something valid makes it render again
PASS document.defaultView.getComputedStyle(rectElement, null).display is "inline"
Check that adding something valid to requiredFeatures keeps rendering the element
PASS document.defaultView.getComputedStyle(rectElement, null).display is "inline"
Check that adding something invalid to requiredFeatures makes it not render
PASS document.defaultView.getComputedStyle(rectElement, null).display is ""
PASS successfullyParsed is true

TEST COMPLETE
